refactor(coding-agent): merge extension API into host package
Amp-Thread-ID: https://ampcode.com/threads/T-019fc8e7-96fe-7678-b712-9544f7d8eb86
