+++ apps/coding-agent/package.json
@@ -82,7 +82,6 @@
     "@ai-sdk/openai-compatible": "3.0.16",
     "@earendil-works/pi-tui": "^0.82.1",
     "@minpeter/opensearch": "0.1.1",
-    "@minpeter/pss-extension-api": "workspace:^",
     "@minpeter/pss-extension-latex": "workspace:^",
     "@minpeter/pss-runtime": "workspace:^",
     "@t3-oss/env-core": "^0.13.11",
