+++ apps/coding-agent/src/extensions/factory-host.test.ts
@@ -1,9 +1,8 @@
 import { createOpenAICompatible } from "@ai-sdk/openai-compatible";
-import { instructions } from "@minpeter/pss-extension-api";
 import { createAgent } from "@minpeter/pss-runtime";
 import { jsonSchema, tool } from "ai";
 import { describe, expect, it } from "vitest";
-import { tools } from "./capabilities";
+import { instructions, tools } from "./capabilities";
 import { createCodingAgentExtensionHost } from "./host";
 import type {
   CodingAgentExtensionApi,
