+++ apps/coding-agent/src/tui/assistant-renderer.ts
@@ -1,9 +1,27 @@
-export type {
-  AssistantRenderer,
-  AssistantRendererContext,
-  AssistantRendererRegistrationOptions,
-  AssistantTextView,
-} from "@minpeter/pss-extension-api";
+import type { Component, MarkdownTheme } from "@earendil-works/pi-tui";
+
+export interface AssistantTextView extends Component {
+  dispose?(): void;
+  setText(text: string): void;
+}
+
+export interface AssistantRendererContext {
+  readonly foregroundColor?: string;
+  readonly markdownTheme: MarkdownTheme;
+  readonly notify: (message: string) => void;
+  readonly notifyOnce: (key: string, message: string) => void;
+  readonly requestRender: () => void;
+  readonly signal: AbortSignal;
+}
+
+export type AssistantRenderer = (
+  context: AssistantRendererContext
+) => AssistantTextView;
+
+export type AssistantRendererRegistrationOptions =
+  | { readonly fallback?: never; readonly override?: never }
+  | { readonly fallback: true; readonly override?: never }
+  | { readonly fallback?: never; readonly override: true };
 
 export interface AssistantRendererNotifications {
   readonly notify: (message: string) => void;
