+++ extensions/latex/package.json
@@ -35,13 +35,13 @@
     "lint": "ultracite check ."
   },
   "dependencies": {
-    "@minpeter/pss-extension-api": "workspace:^",
     "fast-png": "8.0.0",
     "mathjax": "4.1.3",
     "playwright-core": "1.62.0"
   },
   "peerDependencies": {
-    "@earendil-works/pi-tui": "^0.82.1"
+    "@earendil-works/pi-tui": "^0.82.1",
+    "@minpeter/pss-coding-agent": "workspace:^"
   },
   "devDependencies": {
     "tsdown": "^0.22.14",
