+++ extensions/latex/vitest.config.ts
@@ -1,0 +1,28 @@
+import { resolve } from "node:path";
+
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  resolve: {
+    alias: [
+      {
+        find: /^@minpeter\/pss-coding-agent\/extension$/,
+        replacement: resolve(
+          import.meta.dirname,
+          "../../apps/coding-agent/src/extensions/index.ts"
+        ),
+      },
+      {
+        find: /^@minpeter\/pss-runtime$/,
+        replacement: resolve(
+          import.meta.dirname,
+          "../../packages/runtime/src/index.ts"
+        ),
+      },
+    ],
+    conditions: ["@minpeter/pss-source", "import", "module", "default"],
+  },
+  test: {
+    environment: "node",
+  },
+});
