--- packages/extension-api/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@minpeter/pss-extension-api",
-  "version": "0.0.0",
-  "description": "Shared authoring contracts for PSS coding-agent extensions.",
-  "type": "module",
-  "engines": {
-    "node": ">=24"
-  },
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "@minpeter/pss-source": "./src/index.ts",
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/minpeter/pss-runtime.git",
-    "directory": "packages/extension-api"
-  },
-  "publishConfig": {
-    "access": "public",
-    "provenance": true
-  },
-  "scripts": {
-    "build": "tsdown",
-    "typecheck": "tsc -p tsconfig.json --noEmit",
-    "test": "vitest run",
-    "lint": "ultracite check ."
-  },
-  "peerDependencies": {
-    "@earendil-works/pi-tui": "^0.82.1"
-  },
-  "devDependencies": {
-    "tsdown": "^0.22.14",
-    "typescript": "^7.0.2",
-    "vitest": "^4.1.10"
-  }
-}
