--- packages/extension-api/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "customConditions": ["@minpeter/pss-source"],
-    "noEmit": true
-  },
-  "include": ["src/**/*.ts"]
-}
