--- packages/extension-api/tsdown.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "tsdown";
-
-export default defineConfig({
-  dts: true,
-  entry: ["src/index.ts"],
-  fixedExtension: false,
-  root: "src",
-  sourcemap: true,
-  unbundle: true,
-});
